requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
celery>=5.3,<6.0
redis>=5.0,<6.0
python-decouple>=3.8,<4.0
openai>=1.0,<2.0